use engine param with azure openai (#4427)
* use engine param with azure

* fix mypy+isort

GitOrigin-RevId: 80e91dea449dbe747ee817d91ae7d184c6f787c9

+++ examples/pipelines/contextful_s3/app.py
@@ -51,7 +51,7 @@ def run(
     embedding_dimension: int = 1536,
     model_locator: str = "gpt-3.5-turbo",
     max_tokens: int = 60,
-    temperature: int = 0.0,
+    temperature: float = 0.0,
     **kwargs,
 ):
     embedder = OpenAIEmbeddingModel(api_key=api_key)
